+++ client/.eslintignore
@@ -1,3 +1,4 @@
 build/
 node_modules/
 script/
+api/
