+++ client/src/route/dashboard/Dashboard.tsx
@@ -2,60 +2,34 @@
 source: https://github.com/mui/material-ui/tree/v5.10.0/docs/data/material/getting-started/templates/dashboard
 */
 import * as React from 'react';
-import { createTheme, Theme, ThemeProvider } from '@mui/material/styles';
-import Box from '@mui/material/Box';
-import Toolbar from '@mui/material/Toolbar';
-import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
 import Chart from '../history/Chart';
 import RecentRuns from '../history/RecentRuns';
-import Footer from '../../page/Footer';
-import DTaaSMenu from '../../page/Menu';
-
-const mdTheme: Theme = createTheme();
 
 function DashboardContent() {
   return (
-    <ThemeProvider theme={mdTheme}>
-      <Box sx={{ display: 'flex' }}>
-        <DTaaSMenu />
-        <Box
-          component="main"
+    <Grid container spacing={3}>
+      {/* Chart */}
+      <Grid item xs={12} md={8} lg={9}>
+        <Paper
           sx={{
-            flexGrow: 1,
-            height: '100vh',
-            overflow: 'auto',
+            p: 2,
+            display: 'flex',
+            flexDirection: 'column',
+            height: 240,
           }}
         >
-          <Toolbar />
-          <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-            <Grid container spacing={3}>
-              {/* Chart */}
-              <Grid item xs={12} md={8} lg={9}>
-                <Paper
-                  sx={{
-                    p: 2,
-                    display: 'flex',
-                    flexDirection: 'column',
-                    height: 240,
-                  }}
-                >
-                  <Chart />
-                </Paper>
-              </Grid>
-              {/* Past Runs */}
-              <Grid item xs={12}>
-                <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
-                  <RecentRuns />
-                </Paper>
-              </Grid>
-            </Grid>
-            <Footer />
-          </Container>
-        </Box>
-      </Box>
-    </ThemeProvider>
+          <Chart />
+        </Paper>
+      </Grid>
+      {/* Past Runs */}
+      <Grid item xs={12}>
+        <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
+          <RecentRuns />
+        </Paper>
+      </Grid>
+    </Grid>
   );
 }
