+++ client/src/route/digitaltwins/DigitalTwins.tsx
@@ -1,57 +1,28 @@
 import * as React from 'react';
-import { createTheme, Theme, ThemeProvider } from '@mui/material/styles';
-import Box from '@mui/material/Box';
-import Toolbar from '@mui/material/Toolbar';
-import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
-import Footer from '../../page/Footer';
-import DTaaSMenu from '../../page/Menu';
 import Workflows from './Workflows';
-
-const mdTheme: Theme = createTheme();
 
 function DTContent() {
   return (
-    /* jshint ignore:start */
-    <ThemeProvider theme={mdTheme}>
-      <Box sx={{ display: 'flex' }}>
-        <DTaaSMenu />
-        <Box
-          component="main"
+    <Grid container spacing={3}>
+      {/* Components */}
+      <Grid item xs={12} md={8} lg={9}>
+        <Paper
           sx={{
-            flexGrow: 1,
-            height: '100vh',
-            overflow: 'auto',
+            p: 2,
+            display: 'flex',
+            flexDirection: 'column',
+            height: 'fit-content',
           }}
         >
-          <Toolbar />
-          <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-            <Grid container spacing={3}>
-              {/* Components */}
-              <Grid item xs={12} md={8} lg={9}>
-                <Paper
-                  sx={{
-                    p: 2,
-                    display: 'flex',
-                    flexDirection: 'column',
-                    height: 'fit-content',
-                  }}
-                >
-                  <Workflows />
-                </Paper>
-              </Grid>
-            </Grid>
-
-            <Footer />
-          </Container>
-        </Box>
-      </Box>
-    </ThemeProvider>
-    /* jshint ignore:end */
+          <Workflows />
+        </Paper>
+      </Grid>
+    </Grid>
   );
 }
 
 export default function DigitalTwins() {
-  return <DTContent />; /* jshint ignore:line */
+  return <DTContent />;
 }
