+++ client/src/route/history/History.tsx
@@ -1,51 +1,25 @@
 import * as React from 'react';
-import { createTheme, Theme, ThemeProvider } from '@mui/material/styles';
-import Box from '@mui/material/Box';
-import Toolbar from '@mui/material/Toolbar';
-import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
 import RecentRuns from './RecentRuns';
-import Footer from '../../page/Footer';
-import DTaaSMenu from '../../page/Menu';
 import Logs from './Logs';
-
-const mdTheme: Theme = createTheme();
 
 function HistoryContent() {
   return (
-    <ThemeProvider theme={mdTheme}>
-      <Box sx={{ display: 'flex' }}>
-        <DTaaSMenu />
-        <Box
-          component="main"
-          sx={{
-            flexGrow: 1,
-            height: '100vh',
-            overflow: 'auto',
-          }}
-        >
-          <Toolbar />
-          <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-            <Grid container spacing={3}>
-              {/* Past Runs */}
-              <Grid item xs={12}>
-                <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
-                  <RecentRuns />
-                </Paper>
-              </Grid>
-              {/* Logs */}
-              <Grid item xs={12}>
-                <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
-                  <Logs />
-                </Paper>
-              </Grid>
-            </Grid>
-            <Footer />
-          </Container>
-        </Box>
-      </Box>
-    </ThemeProvider>
+    <Grid container spacing={3}>
+      {/* Past Runs */}
+      <Grid item xs={12}>
+        <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
+          <RecentRuns />
+        </Paper>
+      </Grid>
+      {/* Logs */}
+      <Grid item xs={12}>
+        <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
+          <Logs />
+        </Paper>
+      </Grid>
+    </Grid>
   );
 }
